requirements.txt
fastapi
uvicorn
pytesseract
PyPDF2
python-docx
pillow
scikit-learn
numpy
pandas